chore: ResultTest제거 및 기존 테스트 코드 Cards로 이동
